[front] - refactor: standardize chart component imports across the application
 - Moved common chart-related components such as `ChartContainer`, `ChartLegend`, `ChartTooltip`, and `ChartShapes` out of the `agent_builder/observability` directory
 - Updated all import statements to the new shared `charts` directory to maintain consistent usage across different modules
